:package:chore(project): 접근권한 관리를 위해 패키지 구조 도메인 기반으로 변경
