+++ src/main/java/subway/controller/Controller.java
@@ -2,5 +2,5 @@ package subway.controller;
 
 public abstract class Controller {
 
-    abstract void execute();
+    protected abstract void execute();
 }
