+++ src/main/java/subway/domain/line/LineController.java
@@ -1,4 +1,4 @@
-package subway.controller;
+package subway.domain.line;
 
 import static subway.controller.RunStatus.RUNNING;
 import static subway.controller.RunStatus.STOPPED;
@@ -11,9 +11,12 @@ import static subway.view.constants.OutputMessage.LINE_DELETE_INFO;
 import static subway.view.constants.menu.SubCommand.BACK;
 
 import java.util.List;
+import subway.controller.ErrorInterceptor;
+import subway.controller.HandlerAdaptor;
+import subway.controller.ManageController;
+import subway.controller.RunStatus;
 import subway.dto.FinalStationsDTO;
 import subway.dto.LineDTO;
-import subway.service.LineService;
 import subway.view.InputView;
 import subway.view.OutputView;
 import subway.view.constants.menu.SubCommand;
@@ -33,7 +36,7 @@ public class LineController extends ManageController {
     }
 
     @Override
-    void execute() {
+    protected void execute() {
         RunStatus runStatus = RUNNING;
         while (runStatus == RUNNING) {
             OutputView.printLineMenus();
@@ -51,7 +54,7 @@ public class LineController extends ManageController {
     }
 
     @Override
-    void create() {
+    protected void create() {
         String lineName = InputView.inputName(LINE_CREATE_NAME_HEADER);
         String upFinalStationName = InputView.inputName(LINE_CREATE_UP_FINAL_NAME_HEADER);
         String downFinalStationName = InputView.inputName(LINE_CREATE_DOWN_FINAL_NAME_HEADER);
@@ -60,14 +63,14 @@ public class LineController extends ManageController {
     }
 
     @Override
-    void delete() {
+    protected void delete() {
         String lineName = InputView.inputName(LINE_DELETE_HEADER);
         lineService.deleteLine(new LineDTO(lineName));
         OutputView.printInfoMessage(LINE_DELETE_INFO);
     }
 
     @Override
-    void read() {
+    protected void read() {
         List<LineDTO> lines = lineService.getAllLines();
         OutputView.printLines(lines);
     }
