+++ src/main/java/subway/domain/station/StationController.java
@@ -1,4 +1,4 @@
-package subway.controller;
+package subway.domain.station;
 
 import static subway.controller.RunStatus.RUNNING;
 import static subway.controller.RunStatus.STOPPED;
@@ -9,8 +9,11 @@ import static subway.view.constants.OutputMessage.STATION_DELETE_INFO;
 import static subway.view.constants.menu.SubCommand.BACK;
 
 import java.util.List;
+import subway.controller.ErrorInterceptor;
+import subway.controller.HandlerAdaptor;
+import subway.controller.ManageController;
+import subway.controller.RunStatus;
 import subway.dto.StationDTO;
-import subway.service.StationService;
 import subway.view.InputView;
 import subway.view.OutputView;
 import subway.view.constants.menu.SubCommand;
@@ -30,7 +33,7 @@ public class StationController extends ManageController {
     }
 
     @Override
-    public void execute() {
+    protected void execute() {
         RunStatus runStatus = RUNNING;
         while (runStatus == RUNNING) {
             OutputView.printStationMenus();
@@ -48,21 +51,21 @@ public class StationController extends ManageController {
     }
 
     @Override
-    void create() {
+    protected void create() {
         String name = InputView.inputName(STATION_CREATE_NAME_HEADER);
         stationService.addStation(new StationDTO(name));
         OutputView.printInfoMessage(STATION_CREATE_INFO);
     }
 
     @Override
-    void delete() {
+    protected void delete() {
         String name = InputView.inputName(STATION_DELETE_NAME_HEADER);
         stationService.deleteStation(new StationDTO(name));
         OutputView.printInfoMessage(STATION_DELETE_INFO);
     }
 
     @Override
-    void read() {
+    protected void read() {
         List<StationDTO> stationDTOs = stationService.getAllStations();
         OutputView.printStations(stationDTOs);
     }
